pyproject.toml
[project]
name = "mastering-langchain"
version = "0.1.0"
description = "A basic example of LangChain usage"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.0",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
    "streamlit>=1.51.0",
]
